Document the batch package. Separate out the batch tile from the batch queue
